requirements.txt.txt
fastapi
uvicorn
twilio
pandas
apscheduler
python-dotenv
gtts
openai-whisper